% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_triple.R
\name{calculate_triples}
\alias{calculate_triples}
\title{Calculate triples given a clean_date}
\usage{
calculate_triples(clean_date)
}
\arguments{
\item{clean_date}{Date}
}
\value{
data.table
}
\description{
When entered a `date`, calculate the last triple and the next triple from the 
time-distance between today and `date`, that is a combination t of same 
numbers for (month, week, day), for example (1, 1, 1) for the time-distance 
from `date` until today of 1 month, 1 week and 1 day.
}
\examples{
result <- calculate_triple(lubridate::ymd("2020-02-20"))
}
