% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_choose_substance.R
\name{mod_choose_substance_ui}
\alias{mod_choose_substance_ui}
\alias{mod_choose_substance_server}
\title{mod_choose_substance_ui and mod_choose_substance_server}
\usage{
mod_choose_substance_ui(id)

mod_choose_substance_server(input, output, session, sober_time)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
