% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_date_statistic.R
\name{mod_display_date_statistic_ui}
\alias{mod_display_date_statistic_ui}
\alias{mod_display_date_statistic_server}
\title{mod_display_date_statistic_ui and mod_display_date_statistic_server}
\usage{
mod_display_date_statistic_ui(id)

mod_display_date_statistic_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to display a selected date and related statistic
}
\keyword{internal}
