% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input_date.R
\name{mod_input_date_ui}
\alias{mod_input_date_ui}
\alias{mod_input_date_server}
\title{mod_input_date_ui and mod_input_date_server}
\usage{
mod_input_date_ui(id)

mod_input_date_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module to input a date using a date-selector
}
\keyword{internal}
